Fixed trace execution bug
Fixed a bug where a prerecorded trace would not be executed anymore once a loop was restarted after a guard failure because the wrong tracer-context had been passed around in the guard instruction. This guard instruction now uses a global variable containing the tracer-context. This variable is always kept up-to-date.