Transformed the trace
The trace is now a letrec with a single binding for a nullary function which contains all instructions in its body, followed by a tail-recursive call to itself. The body of the letrec just consists of a call to this function.
Executing the trace can be done by simply calling (eval trace).