Initial commit
When executing label-traces, the call/cc stack is no longer used to jump to the next continuation. Instead, it just returns.
When a guard fails, all trace-frames on the execution stack that are an ancestor of the trace that failed are removed and the next continuation is called.
Temporarily commented-out trace merging to simplify debugging