Bugfix
When a merge-point is encountered while tracing, a reference is placed in the current trace to jump to the trace of the tail of this merge-point. However, if the trace for this merge-point-tail is later aborted, e.g., because the trace has become too large, this reference would no longer be valid. This is now solved by checking whether the trace for the merge-point-tail is indeed present and if not, simply continuing with regular interpretation.